pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = ""
authors = ["VictorAnthony506 <victor0anthony@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gunicorn = "^20.1.0"
django = "^4.1.2"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
whitenoise = {extras = ["brotli"], version = "^6.2.0"}
openai = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
